Remove bundled BusyBox binaries
Starting November 1st, 2025, all new apps and updates to existing apps
submitted to Google Play and targeting Android 15+ devices must
support 16 KB page sizes.
BusyBox binaries bundled with libsu violate this requirements.

It should be possible to rebuild BusyBox with the required alignment,
but it's not a trivial task. It looks like most of Magisk's patches
need to be ported to make it work properly on Android. Let's make it a
backup plan. For now, using ToyBox should work fine on most devices.